[s3_env] optimize s3 env backup process
S3 does not support renaming natively, so previously we need to upload these
tmp files into S3, copy them to target name, and then delete these tmp files, so
there are 3 file operations(2 copy + 1 delete). In the new logic, we do the
renaming locally and then upload them to s3, and only 1 copy is needed(local
linux rename is a mv, the overhead could be omitted).