Use rocksdb checkpoint for backup
Checkpoint is a feature in RocksDB which provides the ability to take a snapshot
of a running RocksDB database in a separate directory. And unlike the previous
s3 backup solution, while we need to create a backup to local indirectly(copy sst files and
wal logs) and then upload to s3, due to s3 is not a pure file system as hdfs.
For the checkpoint solutin, it only create the hard links in local and then
upload to s3, which will save us an unnessary copy. We also implemented parallel
upload/download in this approach.