Add state transition handlers for LeaderFollower state model.
This diff adds the state transtion handlers for the LeaderFollower
state model. It's an exact copy of the MasterSlaveStateModelFactory
except that the state names have been replaced to adhere to inclusive
terminology guidelines.

This diff does not change the names of the "Role"s used in
rocksplicator. These will be done in the future. This diff also retains
the "M" and "S" indicators used in the shard map.

Testing: I deployed these changes in a binary and verified that I was
able to successfully initialize a new resource in the LeaderFollower
state model.