fix spectator hosts continue publish config maps even lost lock
In some scenarios where the leader spectator lost the zk connection, the
interprocess lock is released but the it continues to publish config map and the
new leader will also start to generate config map concurrently. We made the fix
by terminating the original leader process when a zk LOST event happened.