Update shard map before notifying followers of new leader.
We have seen issues where a follower has its upstream set to a previous
leader. This causes the follower to not receive any updates from the
current leader, leading to inconsistent data being read by clients.

One of the causes for the above could be that the shard map is
potentially not updated for 10 minutes before notifying the followers.
Although there is a distributed lock for each partition and it is
unlikely that a new follower might have come up while a follower->leader
transition is taking place, considering zookeeper latencies, this is an
easy fix to mitigate the issue to a certain extent.

I was unable to reproduce the issue on a test cluster. So will need to
test it once I land it.