Initial thrift structs for profiling leadership handoff to maintain leader events history (#420)
We want to measure latencies for given resources/clusters in handoff leadership, in order to provide metrics we can track to improve availability of our systems. Helix provides some stats that are useful in understanding avg delays in leadership handoff. However from availability perspective this is not sufficient, as there are some systems that are not part of helix eco-system, but are part of rocksplicator framework. e.g.. Spectator receives notifications for externalView and publishes shard_map to specified url. The client's then relies on the shard_map generated by Spectator for latest routing information. The latencies introduced by Spectator itself in generating the shard_map, as well as the latencies incurred by the external system to which the shard_map is posted by Spectator in making it available to it's clients are not measured... From the client's perspective the system remain unavailable not only during the time when the leader handoff is completed, but also when the client's is actually informed about the new leader. Due to these limitations that helix can't provide in rocksplicator framework, we are building e2e latency measurement and also identifying the parts of the system introducing delays in leader handoff propagation. This is an effort to help identify sub-systems and provide us with direction as to where improvements are needed in order to improve overall availability of the system during deploys or other rebalance activities.

The manner in which we are proposing to do this is by building an event history for each of the partition and maintain such event history in rolling window fashion, ensuring to keep enough events to be able to measure one-full leader handoff cycle.

We will store this information in provided zk cluster (which can be separate from helix zk cluster)...