[Profiling Leader Handoff]: Implementation of zkstore and async read-modiy-write of LeaderEvents. (#423)
LeaderEventsHistoryStore: It encapsulates a zkClient and creates a facade that provides async update through use of executor service. It internally manages instances of zkStore corresponding to each of the resource/partition. It also provides mechanism for pruning out any LeaderEvents that are deemed / guaranteed to be duplicate using a cache of localhistory that is up-to-date w.r.t. changes committed by the local process. It makes sure that only state changes are logged and stored in zk.