[Profiling Leader Handoff]: Implement Config / FileWatcher through polling (#429)
FileWatcher and ConfigStore implementation: We need a mechanism to dynamically specify which resources to log leader events for. In order to do that, we implement a very simple polling on a config file which can be provided as either json array or line separated string values specifying resources to watch.