[replication errors] Add a script to get the current leader of a partition.
This is part 1 of the changes to fix the replication errors we are seeing in read/write clusters.
Brief description of how we are planning to solve it:
When a replication error is seen in rocksdb_replicator, we will query helix to get the latest leader
the partition, and reset the upstream to this latest leader.

This change contains a java script to return the IP of the leader's replica.
I also added a main function to make it easier for standalone testing. But we will be eventually
using the `getLeaderHostIP` function.

The partition_name currently includes the resource name too (since that's how helix stores it)
I will take a call later if it's cleaner to just take in the partition number.

Testing:
Ran the script with various input (and error) combinations:
1. Verified that the expected response is got for good cases.
2. Verified that an empty string was returned for all other cases (
invalid zk, cluster, resource, partition)