[rocksdb_replicator] Add a new thrift field to identify when a leader db is moved.
We use the error code `SOURCE_NOT_FOUND` in two cases:
1. An instance has no information about the db in the replicate request.
2. An instance identifies it is no longer hostting the db requested.

Although the difference is subtle, we want to handle these two
situations differently: we would like to reset the leader only in the
first case (since a new leader might not yet be available in the second
case).

To handle this difference, I'm adding a new field to the `ErrorCode`
struct. Using the new field will be part of a subsequent PR.