[rocksdb_replicator] Reset upstream when the replication error is SOURCE_NOT_FOUND
This is the last diff in the task to fix replication errors on
followers. In this diff, we use the previously implemeneted function to
get the latest upstream for a follower.

Some specifics:
1. Since we do not want to overload the ZX server for external views,
only a small (right now 10) percentage of errors are querying helix.
In the rocksdb_replicator, there is already a delay in the next pull if
there is a replication error (default delay of 5 seconds). So the risk
of overloading ZK is low.
2. Add new a type of replication error: SOURCE_REMOVED. Prior to this
change, if a leader instance was aware that it was no longer hosting a
DB, it would return the same SOURCE_NOT_FOUND error. We want to
differentiate this case to further reduce load on ZK: SOURCE_REMOVED
errors are not eligible for upstream resets since a new leader might not
have been elected yet. We will rely on SOURCE_NOT_FOUND errors for
querying ZK.
3. Adding new gflags for zookeeper server string helix cluster name.
Instead of passing down the flags through a series of constructors, I'm
creating two gflags to set these values. In the service code, I will set
the values from glfags of the service. So, users of the service will be
agnostic of the new gflags.