Catch up to leader before completing offline -> follower transition.
Current behavior in offline -> follower transition for a new offline
shard:
1. Find current leader.
2. Issue backup of DB.
3. Restore DB
4. Done.

Steps 2 and 3 can take up to an hour (and for some cases even more). So
the new follower could be lagging behind the leader for hours worth of
realtime updates.

Helix sometimes promotes a newly provisioned follower immediately to a
leader for load balancing. This causes the follower -> leader transition
to wait until the follower catches up to all the real time updates it
does not have. This causes drop in availability since there is no leader
for this shard during this period of time.

Proposed solution and behavior:
1. Find current leader.
2. Issue backup of DB.
3. Restore DB
4. Wait until the new follower has caught up to the leader (with a small
threshold)
5. Done.

With the proposed solution, future follower -> leader transitions can
complete quicker, improving availability.