Do not crash the process if kafka serverset path is invalid/not present.
If the kafka serverset path is not present, the CHECK would fail and
crash the process. CHECKs should be reserved only during process start
up, not during state transitions.
So removing the use of CHECK in favor of raising an exception. With
this, an exception is returned and the state transition would fail.

Testing:
On a test cluster, sent a startMessageIngestion request with an invalid
serverset path and verified that an error is returned in the API.
```
AdminException: AdminException(message='Failed to start kafka watcher: testtest00016 Failed to add file watcher to /tmp/non-existant-file', errorCode=5)
```