Return a rocksdb status object instead of throwing the return code in 2-ack mode.
When 2-ack mode is enabled and none of the followers ack the write
within the timeout period, a rocksplicator status is thrown (not sure
why). This causes cryptic messages such as:
"worker task threw unhandled non-exception object" in the logs which are
hard to understand.
Replacing that with a return of the rocksdb status. It's up to the
service logic to handle the status. Also logging the error for easier
debugging.

Testing: Forced the issue locally and verified that no exceptions are
thrown, and that the status is returned instead.