[LogExtractor] Ignore log messages which are not timestamps.
With the CDC feature, we will be using Rocksdb's log functionality to
pass metadata to the notifications. So, all log messages are not
timestamps. This change checks the size of the log message to determine
if its a timestamp. There could be a case where the log message size
does match, but is a different type. For now, we will ignore this since
the timestamp is used only for stats and this case is unlikely.

If the log message isn't the size of an unsigned int (8 bytes), the
timestamp is set to 0 so it can be ignored in the places it is read.
I've verified that this check is plresent in all the places it is read.