rocksdb_replicator: reset upstream when fetching from a non-leader that has no updates (#591)
We see in production where writes can timeout after 2-ack mode is enabled, which usually is caused by followers talking to the wrong upstream (typically a follower), hence never ACKing the write. See this internal doc for detailed investigation and plans on fixing those issues.

This PR serves as the initial step to mitigate the issue: trigger a upstream reset when both of the two conditions below are met:

no updates from upstream, for a consecutive number of times (controlled by a GFLAG, in case it's transient)
upstream informs the follower that the upstream is NOT a leader (avoids unnecessary reset when there is no traffic). We introduced the ReplicaRole in Define ReplicaRole in thrift and replace DBRole #592
This new behavior is disabled by default (via a GFLAG reset_upstream_on_empty_updates_from_non_leader) for a safe rollout, but can be reversed once this feature goes out well in production.

Added unit tests for the two edge cases we see in prod (followers setting each other as upstream; and follower setting itself as upstream).