rocksdb_replicator: detect WAL expiration in replication (#599)
We current set WAL TTL to be 1h which has a risk for data loss. On the leader side,
if the requested sequence number do not exist (WAL deletion), it simply returns the
WAL update with the earliest available sequence number.

This PR adds a metrics & log when the above case happens.