rocksdb_replicator: enable leader behind follower stats (#597)
Noticed in prod that we may see leader behind follower sequence, so I wonder if we have metrics for it. Turns out we do, but never enabled it.

Removing the GFLAG since I don't think it's necessary, it's useful to always have this visibility to catch issues: an upstream/leader received a replication request, but turns out its own local sequence number is lower than requested.

Test Plan: deployed a private build and don't see any such metrics emitted, as expected. I don't think this case is expected to happen in production, but if they do, it'll be a good chance to debug why