cluster_management: add simple retry for helix external view (#598)
Noticed in prod where state transitions could fail (thus leading to ERROR shards) due to failure to obtain Helix external view, hence getting NPE when accessing the null view.

It's not clear whether the failure was transient, but I think it's a good practice to add some retry when talking to external dependency. (currently configured to have 3 retries, with 1s, 2s, and 3s sleep in between).

This change affects both LeaderFollower and MasterSlave state models

I plan to have a a follow-up change on dealing with the null view if retry still fails.