rocksdb_replicator: adaptive replication timeout for 2-ACK mode (#610)
When 2-ACK Mode is enabled, writes to the leader will need to wait for at least one follower ACK, before returning to the client. The wait timeout is controlled by a GFLAG replicator_timeout_ms, which can be generous to accommodate for the p99/p999 network delay (e.g. we typically set it to be 1s in production). However, in the cases when there is replication issues for a db, the leader may not be receiving followers' ACK for a long time, and as a result, all writes to this db would be blocked for 1s, which may end up blocking all worker threads and impacting all other dbs in the process. In production, we see this happens quite often, which leads to load shedding errors in the thrift handler and other cascading failures (e.g. state transitions may fail due to this).

As a mitigation, we introduce adaptive replication timeout, i.e.:

- We keep the current GFLAG_replication_timeout_ms as the normal timeout (e.g. 1s)
- We introduce a new GFLAG_replication_timeout_degraded_ms which can be very small (e.g. 10ms), based on stats: https://statsboard.pinadmin.com/share/j7qev, p99 is usually around 1ms, so requests should typically still succeed in that case.
- For a DB, if we see N (e.g. N=100) consecutive timeouts, we know that it's likely there is replication issue and the leader cannot receive follower ACK, so we fall back to the degraded mode, and use GFLAG_replication_timeout_degraded_ms instead. The goal is to fail-fast so that we can yield resource for other DBs.
- In degradation mode, as soon as there is one successful request (i.e. follower ACK returned with GFLAG_replication_timeout_degraded_ms), we immediately recover to normal mode with the regular GFLAG_replication_timeout_ms.

Also, while we are here, updating the default replicator_timeout_ms GFLAG to be 2s based on our production stats that should be sufficient for the worst-case.