LeaderFollowerStateModelFactory: set rebuild WAL lag upper bound (#614)
During FOLLOWER->LEADER state transition, follower may decide not to "rebuild" (restore from upstream snapshot) local db if its latest local db update is within the WAL TTL (default to 1h), i.e. that means newer updates have not been expired/deleted in the upstream, so it's able to catch up from replication RPC.

However, WAL TTL can be arbitrary set, while we default it to 1h, application may increase it to be longer (e.g. 4h, or even a day). In that case, we should probably trigger a rebuild if the follower is too lagging behind, since log catchup takes time.

This PR proposes a simple change: regardless of the WAL TTL, we'll trigger rebuild if the local WAL is stale for more than one hour. We chose 1h as a starting point, because that's what the default WAL TTL is (so no production behavior change to existing applications that use the default WAL TTL), we may tune it in the future (e.g. further shrink it if our backup&restore path becomes more reliable)

ref: https://jira.pinadmin.com/browse/KV-5337