cluster_management: support an option to close admin client connection (#615)
The current getAdminClient creates a Admin thrift client for various usage during state transition or
even during runtime (e.g. in PartitionStateUpdater). Under the hood it opens a TCP socket which the
thrift client uses to establish connections.

The socket is not closed after usage, which could be problematic, if there are a large number of admin clients
used. This is the case for the usage in PartitionStateUpdater, which schedules all leader partitions in a process
to retrieve local db state (via admin client) every few minutes. Eventually, we see process crash or run into failing state where it exhausted the socket count.

This PR proposes a simple fix: provide a new getAdminClientOnSocket to asks the caller to provide the socket
and hence let the caller know they should be responsible for closing the socket.

Test plan:
- mvn test still work
- tested in a private build, among many other tests, please see an internal doc for detailed investigation and validation.