rocksdb_repliator: do not count observer replication ACK (#618)
On replication mode TWO (2-ack mode), upon writes, the leader will need to wait for at least
one follower to receive the update, before returning successfully to to the client. Currently it does so by tracking the max_seq_no_acked_ which is the maximum ACK-ed sequence number from any replication requests. That's problematic, because the replication requests could come from either FOLLOWER or OBSERVER. See #601 where we introduced the OBSERVER role.

This PR corrects that behavior, specifically, on the FOLLOWER/OBSERVER side, set the role in each replication request, and on the leader side, do not count OBSERVER sequence number when updating max_seq_no_acked_. Log a counter for visibility in that case.